change: rename files in Category page
