+++ src/pages/Capacity/Category.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import UniversimeApi from '@/services/UniversimeApi';
 import './Category.css';
-import ContentStar from './Components/Video/VideoStar';
+import ContentStar from './Components/Content/ContentStar';
 import NotFoundVideo from './Components/NotFoundVideo/NotFoundVideo';
 import Footer from '@/components/Footer/Footer';
 import StarRating from './Components/StarRating/StarRating';
