+++ src/pages/Capacity/Components/Content/ContentStar.css
